+++ spherical_array_processing/__init__.py
@@ -102,7 +102,7 @@ from __future__ import annotations
 
 from importlib import import_module
 
-__version__ = "0.4.0"
+__version__ = "0.5.0"
 
 from .types import (
     ArrayGeometry,